+++ src/config.rs
@@ -96,19 +96,13 @@ pub fn write_default_file_if_missing() {
             inactive: (BooleanLightMode::On, RedAmberGreenLightMode::Green),
             active: (BooleanLightMode::On, RedAmberGreenLightMode::Amber),
             blocked: (BooleanLightMode::Off, RedAmberGreenLightMode::Red),
-            alert: (
-                BooleanLightMode::Flash,
-                RedAmberGreenLightMode::RedAmberFlash,
-            ),
+            alert: (BooleanLightMode::Flash, RedAmberGreenLightMode::AmberFlash),
         },
         hardpoints_deployed: Some(ModeConfig {
             inactive: (BooleanLightMode::On, RedAmberGreenLightMode::Red),
             active: (BooleanLightMode::On, RedAmberGreenLightMode::Amber),
             blocked: (BooleanLightMode::Off, RedAmberGreenLightMode::Off),
-            alert: (
-                BooleanLightMode::Flash,
-                RedAmberGreenLightMode::RedAmberFlash,
-            ),
+            alert: (BooleanLightMode::Flash, RedAmberGreenLightMode::AmberFlash),
         }),
         night_vision: None,
     };